[Fix] 개발 환경 세팅

+++ src/main/java/com/writon/admin/domain/repository/challenge/ChallengeDayRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.challenge;
 import com.writon.admin.domain.entity.challenge.ChallengeDay;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeDayRepository extends JpaRepository<ChallengeDay, Integer> {
+public interface ChallengeDayRepository extends JpaRepository<ChallengeDay, Long> {
 
 }
