+++ src/main/java/com/writon/admin/domain/repository/challenge/ChallengeDepositDeductionRepository.java
@@ -4,6 +4,6 @@ import com.writon.admin.domain.entity.challenge.ChallengeDepositDeduction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeDepositDeductionRepository extends
-    JpaRepository<ChallengeDepositDeduction, Integer> {
+    JpaRepository<ChallengeDepositDeduction, Long> {
 
 }
