+++ src/main/java/com/writon/admin/domain/repository/challenge/ChallengeRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.challenge;
 import com.writon.admin.domain.entity.challenge.Challenge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
 
 }
