+++ src/main/java/com/writon/admin/domain/repository/challenge/EmailRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.challenge;
 import com.writon.admin.domain.entity.challenge.Email;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailRepository extends JpaRepository<Email, Integer> {
+public interface EmailRepository extends JpaRepository<Email, Long> {
 
 }
