+++ src/main/java/com/writon/admin/domain/repository/challenge/FunctionRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.challenge;
 import com.writon.admin.domain.entity.challenge.Function;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FunctionRepository extends JpaRepository<Function, Integer> {
+public interface FunctionRepository extends JpaRepository<Function, Long> {
 
 }
